Move the upscaling rule computation into its file
